return xlate
